Moving github related methods to their own file and creating skeleton for Controller-based repo organization
